requirements.txt
esptool==5.0.1
QPUIQ
pyserial
PySide6_Essentials
pyinstaller
